requirements.txt
pytesseract==0.3.13
Pillow==12.1.1
streamlit==1.55.0
numpy==2.4.2
pandas==2.3.3
matplotlib==3.10.8
scikit-learn==1.8.0